Create README.md
Problem Statement - The data contains 120 years of 
Olympics history. 
The ultimate goal of the project is to identify the key 
insights of the data that can help the management to make 
data driven decisions for the upcoming Olympics event. 
Approach - Created an ER- Diagram(schema) for the 
database. 
Conducted queries and analysis to retrieve useful 
information and insights from the database, 
demonstrating proficiency in SQL and data manipulation 
techniques such as joins, aggregations and by using 
various MYSQL functions. 
Applied problem-solving and critical thinking skills to 
troubleshoot and debug any issues that arose during the